Table, RealtimeTable components are extracted to Sticker component.
